backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]>=3.1.0
geoalchemy2==0.14.2
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
polyline==2.0.0
alembic==1.13.1
openai==1.3.0
pillow>=10.2.0
boto3>=1.28.0

